Reorganize widgets into directories for each 'view.'
