+++ src/find_missing_words/gui/utils/note_field_tree.py
@@ -6,7 +6,7 @@ See docs/data-structures/note_field_tree.md for details on the data structure.
 from aqt import mw
 from aqt.qt import *
 
-from .forms import note_field_tree as tree_form
+from ..forms import note_field_tree as tree_form
 
 
 class NoteFieldTree(QDialog):
